linkedlist: Added iterator functions and tests.
Linked lists can now be manually iterated over which may provides a performance
boost over manually iterating with linkedlist_get. Iterator functions exist for
both forwards and backwards iterating and can be iterated back and forth
because the iterator is a macro for node_s and the linked list is implemented
as a double linked list.